JS-7559: invite link logic
