+++ src/scss/menu/select.scss
@@ -113,4 +113,19 @@
 			.icon.reply { background-image: url('~img/icon/chat/buttons/reply0.svg'); }
 		}
 	}
+
+	/* Invite Link */
+	.menu.menuSelect.inviteLinkType { width: var(--menu-width-large); }
+	.menu.menuSelect.inviteLinkType {
+		.item {
+			.icon {
+				width: 40px; height: 40px; border-radius: 50%; background-size: 20px;
+				background-position: 50% 50%; background-color: var(--color-shape-highlight-medium);
+			}
+			.icon.editor { background-image: url('~img/icon/menu/inviteLink/editor.svg'); }
+			.icon.viewer { background-image: url('~img/icon/menu/inviteLink/viewer.svg'); }
+			.icon.manual { background-image: url('~img/icon/menu/inviteLink/manual.svg'); }
+			.icon.none { background-image: url('~img/icon/menu/inviteLink/none.svg'); }
+		}
+	}
 }
