+++ src/ts/component/page/main/settings/space/share/members.tsx
@@ -1,0 +1,360 @@
+import * as React from 'react';
+import { observer } from 'mobx-react';
+import { Title, Label, Icon, Button, IconObject, ObjectName, Tag, Loader } from 'Component';
+import { I, C, S, U, translate, Action, analytics, } from 'Lib';
+import { AutoSizer, WindowScroller, CellMeasurer, CellMeasurerCache, List } from 'react-virtualized';
+
+interface State {
+	isLoading: boolean;
+};
+
+const HEIGHT = 64;
+
+const Members = observer(class Members extends React.Component<I.PageSettingsComponent, State> {
+
+	node: any = null;
+	cache: any = null;
+	top = 0;
+	refList: any = null;
+
+	state = {
+		isLoading: false,
+	};
+
+	constructor (props: I.PageSettingsComponent) {
+		super(props);
+
+		this.onScroll = this.onScroll.bind(this);
+		this.onChangePermissions = this.onChangePermissions.bind(this);
+		this.onPermissionsSelect = this.onPermissionsSelect.bind(this);
+		this.onUpgrade = this.onUpgrade.bind(this);
+	};
+
+	render () {
+		const { isLoading } = this.state;
+
+		if (isLoading) {
+			return <Loader id="loader" />;
+		};
+
+		const { membership } = S.Auth;
+		const tier = U.Data.getMembershipTier(membership.tier);
+		const space = U.Space.getSpaceview();
+		const participant = U.Space.getParticipant();
+		const members = this.getParticipantList();
+		const length = members.length;
+		const isSpaceOwner = U.Space.isMyOwner();
+
+		let limitLabel = '';
+		let limitButton = '';
+		let showLimit = false;
+		let memberUpgradeType = '';
+
+		if (space.isShared) {
+			if (!U.Space.getReaderLimit() && tier?.price) {
+				limitLabel = translate('popupSettingsSpaceShareInvitesReaderLimitReachedLabel');
+				limitButton = translate('popupSettingsSpaceShareInvitesReaderLimitReachedButton');
+				memberUpgradeType = 'members';
+				showLimit = true;
+			} else
+			if (!U.Space.getWriterLimit()) {
+				limitLabel = translate('popupSettingsSpaceShareInvitesWriterLimitReachedLabel');
+				limitButton = translate('popupSettingsSpaceShareInvitesWriterLimitReachedButton');
+				memberUpgradeType = 'editors';
+				showLimit = true;
+			};
+		};
+
+		const Member = (item: any) => {
+			const isCurrent = item.id == participant?.id;
+
+			let tag = null;
+			let button = null;
+
+			if (isSpaceOwner) {
+				if (isCurrent) {
+					button = <Label text={translate(`participantPermissions${item.permissions}`)} />;
+				} else
+				if (item.isJoining) {
+					button = (
+						<Button
+							className="c36"
+							color="blank"
+							text={translate('popupSettingsSpaceShareViewRequest')}
+							onClick={() => this.onJoinRequest(item)}
+						/>
+					);
+				} else
+				if (item.isRemoving) {
+					button = (
+						<Button
+							className="c36"
+							color="blank"
+							text={translate('commonApprove')}
+							onClick={() => this.onLeaveRequest(item)}
+						/>
+					);
+				} else {
+					button = (
+						<div id={`item-${item.id}-select`} className="select" onClick={() => this.onPermissionsSelect(item)}>
+							<div className="item">
+								<div className="name">{translate(`participantPermissions${item.permissions}`)}</div>
+							</div>
+							<Icon className="arrow dark" />
+						</div>
+					);
+				};
+			} else
+			if (item.isActive) {
+				button = <Label color="grey" text={translate(`participantPermissions${item.permissions}`)} />;
+			} else
+			if (item.isDeclined || item.isRemoved) {
+				button = <Label color="red" text={translate(`participantStatus${item.status}`)} />;
+			};
+
+			if (item.isJoining) {
+				tag = <Tag text={translate('popupSettingsSpaceShareJoinRequest')} />;
+			} else
+			if (item.isRemoving) {
+				tag = <Tag text={translate('popupSettingsSpaceShareLeaveRequest')} />;
+			};
+
+			return (
+				<div id={`item-${item.id}`} className="row" style={item.style} >
+					<div className="side left" onClick={() => U.Object.openConfig(item)}>
+						<IconObject size={48} object={item} />
+						<ObjectName object={item} />
+						{tag}
+						{isCurrent ? <div className="caption">({translate('commonYou')})</div> : ''}
+					</div>
+					<div className="side right">
+						{button}
+					</div>
+				</div>
+			);
+		};
+
+		const rowRenderer = (param: any) => {
+			const item: any = members[param.index];
+			return (
+				<CellMeasurer
+					key={param.key}
+					parent={param.parent}
+					cache={this.cache}
+					columnIndex={0}
+					rowIndex={param.index}
+					hasFixedWidth={() => {}}
+				>
+					<Member key={item.id} {...item} index={param.index} style={param.style} />
+				</CellMeasurer>
+			);
+		};
+
+		return (
+			<div
+				ref={node => this.node = node}
+				id="sectionMembers"
+				className="section sectionMembers"
+			>
+				<div className="membersTitle">
+					<Title text={translate('commonMembers')} />
+					{length > 1 ? <Label text={String(length)} /> : ''}
+				</div>
+
+				{showLimit ? (
+					<div className="row payment">
+						<Label text={limitLabel} />
+						<Button className="payment" text={limitButton} onClick={() => this.onUpgrade(memberUpgradeType)} />
+					</div>
+				) : ''}
+
+				{this.cache ? (
+					<div id="list" className="rows">
+						<WindowScroller scrollElement={window}>
+							{({ height, isScrolling, registerChild, scrollTop }) => (
+								<AutoSizer disableHeight={true} className="scrollArea">
+									{({ width }) => (
+										<List
+											ref={ref => this.refList = ref}
+											autoHeight={true}
+											height={Number(height) || 0}
+											width={Number(width) || 0}
+											deferredMeasurmentCache={this.cache}
+											rowCount={length}
+											rowHeight={HEIGHT}
+											rowRenderer={rowRenderer}
+											onScroll={this.onScroll}
+											isScrolling={isScrolling}
+											scrollTop={scrollTop}
+										/>
+									)}
+								</AutoSizer>
+							)}
+						</WindowScroller>
+					</div>
+				) : ''}
+			</div>
+		);
+	};
+
+	componentDidMount () {
+		const items = this.getParticipantList();
+
+		this.cache = new CellMeasurerCache({
+			fixedWidth: true,
+			defaultHeight: HEIGHT,
+			keyMapper: i => (items[i] || {}).id,
+		});
+
+		this.forceUpdate();
+
+		analytics.event('ScreenSettingsSpaceShare');
+	};
+
+	componentDidUpdate() {
+		this.resize();
+	};
+
+	onScroll ({ scrollTop }) {
+		if (scrollTop) {
+			this.top = scrollTop;
+		};
+	};
+
+	onUpgrade (type: string) {
+		Action.membershipUpgrade();
+
+		analytics.event('ClickUpgradePlanTooltip', { type, route: analytics.route.settingsSpaceShare });
+	};
+
+	getParticipantList () {
+		const records = U.Space.getParticipantsList([ I.ParticipantStatus.Joining, I.ParticipantStatus.Active ]);
+
+		return records.sort((c1, c2) => {
+			const isRequest1 = c1.isJoining || c1.isRemoving;
+			const isRequest2 = c2.isJoining || c2.isRemoving;
+			const cd1 = c1.createdDate;
+			const cd2 = c2.createdDate;
+
+			if (isRequest1 && !isRequest2) return -1;
+			if (!isRequest1 && isRequest2) return 1;
+			if (isRequest1 && isRequest2) return cd1 < cd2 ? -1 : 1;
+
+			return 0;
+		});
+	};
+
+	getParticipantOptions () {
+		const { membership } = S.Auth;
+		const tier = U.Data.getMembershipTier(membership.tier);
+
+		let items: any[] = [] as any[];
+
+		if (!tier?.price || (U.Space.getReaderLimit() - 1 >= 0)) {
+			items.push({ id: I.ParticipantPermissions.Reader });
+		};
+		if (!tier?.price || (U.Space.getWriterLimit() - 1 >= 0)) {
+			items.push({ id: I.ParticipantPermissions.Writer });
+		};
+
+		items = items.map(it => {
+			it.name = translate(`participantPermissions${it.id}`);
+			return it;
+		});
+
+		if (items.length) {
+			items.push({ isDiv: true });
+		};
+
+		items.push({ id: 'remove', name: translate('popupSettingsSpaceShareRemoveMember'), color: 'red' });
+
+		return items;
+	};
+
+	onPermissionsSelect (item: any) {
+		S.Menu.open('select', {
+			element: `#item-${item.id}-select`,
+			horizontal: I.MenuDirection.Right,
+			data: {
+				value: item.permissions,
+				options: this.getParticipantOptions(),
+				onSelect: (e: any, el: any) => {
+					this.onChangePermissions(item, el.id);
+				},
+			},
+		});
+	};
+
+	onChangePermissions (item: any, v: any) {
+		const { space } = S.Common;
+
+		let title = '';
+		let text = '';
+		let button = '';
+		let onConfirm = null;
+
+		switch (v) {
+			case 'remove': {
+				title = translate('popupConfirmMemberRemoveTitle');
+				text = U.Common.sprintf(translate('popupConfirmMemberRemoveText'), item.name);
+				button = translate('commonRemove');
+
+				onConfirm = () => {
+					C.SpaceParticipantRemove(space, [ item.identity ]);
+
+					analytics.event('RemoveSpaceMember');
+				};
+				break;
+			};
+
+			default: {
+				v = Number(v) || I.ParticipantPermissions.Reader;
+
+				title = translate('commonAreYouSure');
+				text = U.Common.sprintf(translate('popupConfirmMemberChangeText'), item.name, translate(`participantPermissions${v}`));
+
+				onConfirm = () => {
+					C.SpaceParticipantPermissionsChange(space, [ { identity: item.identity, permissions: Number(v) } ]);
+
+					analytics.event('ChangeSpaceMemberPermissions', { type: v });
+				};
+				break;
+			};
+		};
+
+		S.Popup.open('confirm', {
+			data: {
+				title,
+				text,
+				textConfirm: button,
+				colorConfirm: 'red',
+				onConfirm,
+			},
+		});
+	};
+
+	onJoinRequest (item: any) {
+		S.Popup.open('inviteConfirm', {
+			data: {
+				name: item.name,
+				icon: item.iconImage,
+				spaceId: S.Common.space,
+				identity: item.identity,
+				route: analytics.route.settings,
+			}
+		});
+	};
+
+	onLeaveRequest (item: any) {
+		Action.leaveApprove(S.Common.space, [ item.identity ], item.name, analytics.route.settings);
+	};
+
+	resize () {
+		if (this.refList) {
+			this.refList.recomputeRowHeights(0);
+		};
+	};
+
+});
+
+export default Members;
