+++ src/ts/lib/api/response.ts
@@ -641,6 +641,7 @@ export const SpaceInviteGetCurrent = (response: Rpc.Space.InviteGetCurrent.Respo
 		inviteCid: response.getInvitecid(),
 		inviteKey: response.getInvitefilekey(),
 		inviteType: response.getInvitetype(),
+		permissions: response.getPermissions(),
 	};
 };
